# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.AddTokensToBotBalanceRequest:
    properties:
      amount:
        description: Сумма для добавления
        type: number
      token_type:
        description: Тип токена (например, ton_balance)
        type: string
    required:
    - amount
    - token_type
    type: object
...
  controllers.InitializeBotBalanceRequest:
    properties:
      dfc_balance:
        description: Баланс DFC
        type: number
      m5_balance:
        description: Баланс M5
        type: number
      ton_balance:
        description: Баланс Ton
        type: number
    required:
    - dfc_balance
    - m5_balance
    - ton_balance
    type: object
  controllers.PlayDiceGameRequest:
    properties:
      bet_amount:
        description: Ставка
        type: number
      target_score:
        description: Цель по очкам
        maximum: 45
        minimum: 15
        type: integer
      token_type:
        description: Тип токена
        type: string
      wallet:
        description: Адрес кошелька
        type: string
    required:
    - bet_amount
    - target_score
    - token_type
    - wallet
    type: object
  controllers.SubtractTokensFromBotBalanceRequest:
    properties:
      amount:
        description: Сумма для вычитания
        type: number
      token_type:
        description: Тип токена (например, ton_balance)
        type: string
    required:
    - amount
    - token_type
    type: object
...
  entities.BotBalanceEntity:
    properties:
      created_at:
        type: string
      dfc_balance:
        type: number
      m5_balance:
        type: number
      ton_balance:
        type: number
      updated_at:
        type: string
    type: object
...
host: api.m5dice.com
info:
  contact: {}
  description: API for managing users and games in TG-Dice application
  title: TG-Dice API
  version: "1.0"
paths:
  /bot/balance:
    get:
      consumes:
      - application/json
      description: Возвращает все балансы бота
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.BotBalanceEntity'
        "500":
          description: Ошибка при получении баланса бота
          schema:
            type: string
      summary: Получает все балансы бота
      tags:
      - bot-balance
  /bot/balance/{tokenType}:
    get:
      consumes:
      - application/json
      description: Возвращает баланс конкретного токена
      parameters:
      - description: Тип токена (ton_balance, m5_balance, dfc_balance)
        in: path
        name: tokenType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: number
        "400":
          description: Некорректный тип токена
          schema:
            type: string
        "500":
          description: Ошибка при получении баланса токена
          schema:
            type: string
      summary: Получает баланс конкретного токена
      tags:
      - bot-balance
  /bot/balance/add:
    post:
      consumes:
      - application/json
      description: Adds the specified amount of tokens to the bot balance
      parameters:
      - description: Token type and amount to add
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.AddTokensToBotBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add tokens to bot balance
      tags:
      - bot-balance
  /bot/balance/subtract:
    post:
      consumes:
      - application/json
      description: Subtracts the specified amount of tokens from the bot balance
      parameters:
      - description: Token type and amount to subtract
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.SubtractTokensFromBotBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Subtract tokens from bot balance
      tags:
      - bot-balance
  /games/bot/balance:
    post:
      consumes:
      - application/json
      description: Initialize bot balance in the system
      parameters:
      - description: Bot balance data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.InitializeBotBalanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Баланс бота успешно инициализирован
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка с параметрами запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при обработке запроса
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Initialize Bot Balance
      tags:
      - bot
  /games/dice:
    post:
      consumes:
      - application/json
      description: Play a dice game with the bot, returning detailed round-by-round
        results
      parameters:
      - description: Game data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.PlayDiceGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Игровой результат
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Ошибка с параметрами запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при обработке запроса
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Play Dice Game with Bot
      tags:
      - bot
      - games
...
  /games/simulate-user-win/{wallet}:
    post:
      consumes:
      - application/json
      description: Симулирует игру, где пользователь выигрывает. Результат игры и
        начисление кубов.
      parameters:
      - description: Wallet адрес пользователя
        in: path
        name: wallet
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'result: game result'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: wallet is required'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: error message'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Симуляция победы пользователя в игре
      tags:
      - games
...
schemes:
- https
swagger: "2.0"